requirements.txt
# This file contains the required PyPi packages for the project
#
# Serial Communication
pyserial
onewire
RPi.GPIO
#
# User Interface
# tkinter
#
# Database
# More to come...

